third_party/blink/web_tests/virtual/xslt-disabled/http/tests/xsl/xslt-transform-with-javascript-disabled-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been removed in this browser. See https://chromestatus.com/feature/4709671889534976.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is the first paragraph This is another paragraph This is the last paragraph
